# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aveline API
  version: "1.0.0"
  description: |
    Aveline's public HTTP API — a boutique concierge platform covering billing,
    entitlements, statistics, users, and organizations.

    ## Versioning
    All user-facing routes are prefixed `/api/v1`. There is no versioning library:
    the prefix is a literal route group. Additive changes do not bump the version;
    clients MUST ignore unknown fields.

    ## Authentication
    Three schemes:
    - `bearerAuth` — Clerk JWT. Default for every user-facing route.
    - `internalToken` — `X-Internal-Token` shared secret. `/internal/**` only. NOT for frontend clients.
    - `apiKeyAuth` — `X-Api-Key`. Customer server-to-server integrations on the Rose plan.

    A route uses exactly one scheme.

    ## Authorization
    Two layers, failing closed:
    1. A named policy for the required permission or role.
    2. For any route containing `{organizationId}`, an `Active` OrganizationMembership
       whose boutique role grants the required permission. A stale JWT cannot cross organizations.

    ## Errors
    A single global exception handler (`GlobalExceptionHandler`, registered as the
    outermost middleware at `Program.cs:99`) turns every unhandled exception into a
    stable `{ status, message, traceId }` `500` body. The shapes are:
    - `400` field validation -> `ValidationProblemDetails`
    - `400`/`404`/`409`/`502`/`503` -> `ErrorResponse` (`{ message }`)
    - `401` -> empty body
...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/orgs/{organizationId}/conversations:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Conversations]
      summary: List conversations
      operationId: listConversations
      description: >-
        The inbox, newest-first with an id tiebreak so paging is total. Page one holds
        the newest threads. Each row carries the client's name, a block-aware preview of
        the newest message, the block it came from (the row's category), who spoke last
        and the actionable `markers` set. The same tile is pushed over the SignalR
        `conversations` hub as **`ReceiveConversationChanged`** whenever a thread changes
        or appears (to `org:{organizationId}` for shared threads and to
        `user:{ownerUserId}` for a per-user general Salon), so an open inbox updates
        without a re-list.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversationPage" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Conversations]
      summary: Create a conversation
      operationId: createConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId: { type: string, format: uuid, nullable: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversationDto" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/orgs/{organizationId}/conversations/{conversationId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/ConversationId"
    get:
      tags: [Conversations]
      summary: Read a conversation
      operationId: getConversation
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversationDto" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{organizationId}/conversations/{conversationId}/messages:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/ConversationId"
    get:
      tags: [Conversations]
      summary: List messages
      operationId: listMessages
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: around
          in: query
          schema: { type: string, format: uuid, nullable: true }
          description: Deep-link anchor; returns a window centred on this message id
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessagePage" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Conversations]
      summary: Send a staff message
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1 }
      responses:
        "200":
          description: Message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageDto" }
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{organizationId}/conversations/{conversationId}/select-customer:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/ConversationId"
    post:
      tags: [Conversations]
      summary: Bind a customer to a conversation
      operationId: selectConversationCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerId]
              properties:
                customerId: { type: string, format: uuid }
                query: { type: string, nullable: true }
      responses:
        "200":
          description: Bound
        "400":
          description: A customerId is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{organizationId}/conversations/{conversationId}/messages/{messageId}/sign-off:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/ConversationId"
      - name: messageId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Conversations]
      summary: Decide a sign-off request
      description: |
        The `contentHash` must match the hash of the message as the approver saw it.
        A mismatch is rejected so an approval cannot be applied to edited content.
      operationId: decideSignOff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved, contentHash]
              properties:
                approved: { type: boolean }
                contentHash: { type: string, minLength: 1 }
      responses:
        "200":
          description: Decided
        "400":
          description: Content hash missing or mismatched
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Integrations
  # ---------------------------------------------------------------------------
...
  /api/v1/orgs/{organizationId}/statistics/api/endpoints:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [API Statistics]
      summary: Read per-endpoint usage
      description: Phase 5. Requires `stats:view`.
      operationId: getApiEndpointStatistics
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - name: sort
          in: query
          schema:
            type: string
            enum: [count, -count, errorRate, -errorRate]
            default: -count
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Endpoint usage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEndpointStatisticsResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Clerk JWT from the `jwt-aveline-v1` template. Validated against the Clerk
        JWKS (signature, `iss`, `exp`, `nbf`). `aud` is deliberately not validated.
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: |
        Service-to-service shared secret. Required on `/internal/**` and `/metrics`.
        Not for frontend clients.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Customer API key, formatted `avl_live_<32 base62 characters>` or
        `avl_test_<...>`. Phase 3. Available on the Rose plan and above.
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: |
        The organization id. Its presence in the path is mandatory for
        organization-scoped authorization: the handler reads only this route value.
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    From:
      name: from
      in: query
      required: true
      schema: { type: string, format: date-time }
      description: ISO 8601 UTC, inclusive
    To:
      name: to
      in: query
      required: true
      schema: { type: string, format: date-time }
      description: ISO 8601 UTC, exclusive
...
  responses:
    Unauthorized:
      description: |
        401. The body is **empty** for JWT failures. Send the client through
        re-authentication.
      content:
        application/json:
          schema:
            description: Usually empty
            nullable: true
    Forbidden:
      description: |
        403. The body is **empty** for policy failures. For the account-state gate
        the body is a `ProblemDetails`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found, or the caller is not a member of the target organization
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: One or more request fields failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationProblemDetails" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code:
          type: string
          description: Machine-readable code. Present on Part C endpoints.
          example: insufficient-balance
        requestId:
          type: string
          description: Correlation id. Phase 0 onwards.
    ValidationProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    PageMeta:
      type: object
      required: [total, page, pageSize]
      properties:
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 200 }
...
    ConversationDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          description: >-
            The thread's nature and audience. `Salon` is the agent-to-staff thread;
            `Announcement` and `Digest` are reserved. Customer context is a separate axis
            carried by `customerId`, and clients classify by that context first.
        customerId: { type: string, format: uuid, nullable: true }
        customerName: { type: string, nullable: true }
        externalRef:
          type: string
          nullable: true
          description: >-
            The channel reference (e.g. a WhatsApp phone number). Present on a
            channel-created thread even when the customer is not yet identified, so the
            client renders it as an unnamed client rather than pinning it as the concierge.
        threadId: { type: string }
        status: { type: string }
        lastMessageAt: { type: string, format: date-time, nullable: true }
        lastMessagePreview:
          type: string
          nullable: true
          description: >-
            A block-aware summary of the newest message, truncated to 120 characters. Null
            when the message carries no block the inbox knows how to summarise.
        lastMessageKind:
          type: string
          nullable: true
          description: The message kind as sent (`Note` for every agent persona).
        lastMessageBlock:
          type: string
          nullable: true
          description: >-
            The content-block type the preview was derived from. This is the row's
            category, because agent output is published as `kind: Note` and the category
            lives in the blocks.
        lastMessageAuthor:
          type: string
          nullable: true
          description: One of `Staff`, `Agent`, `Customer`.
        lastMessageAgentKey: { type: string, nullable: true }
        markers:
          type: array
          description: >-
            The row's actionable states, sorted by fixed priority
            `approval` -> `choice` -> `draft`. A set rather than one value, because a
            thread can hold more than one.
          items:
            type: string
            enum: [approval, choice, draft]
    ConversationPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/ConversationDto" }
    MessageDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        authorKind: { type: string }
        agentKey: { type: string, nullable: true }
        authorUserId: { type: string, format: uuid, nullable: true }
        kind: { type: string }
        contentBlocks:
          type: array
          items: { type: object, additionalProperties: true }
        contentHash: { type: string, nullable: true }
        replyToMessageId: { type: string, format: uuid, nullable: true }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    MessagePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/MessageDto" }

    # --- integrations --------------------------------------------------------
...
    ApiEndpointStatisticsResponse:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  routeTemplate: { type: string }
                  httpMethod: { type: string }
                  requestCount: { type: integer }
                  errorRate: { type: number }
                  avgMs: { type: integer, nullable: true }
...